Splines are created differently, possibly fixing existing bugs
Theory is mostly the same, but instead of inductively defining control points, we
solve a matrix equation for the first control points and then solve for the second
control points with a couple formulas. Overall, this provides a better-looking
and more correct spline.

However, if certain points are close together compared to others, the control points
explode away from them. Looking into fixing that.